+++ lib/rails_template18f/generators.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module RailsTemplate18f
+  module Generators
+    extend ActiveSupport::Autoload
+
+    autoload :Base
+    autoload :CloudGovOptions
+    autoload :PipelineOptions
+  end
+end