+++ lib/rails_template18f/generators/base.rb
@@ -1,0 +1,32 @@
+# frozen_string_literal: true
+
+module RailsTemplate18f
+  module Generators
+    module Base
+      extend ActiveSupport::Concern
+      include ::Rails::Generators::AppName
+
+      included do
+        self.source_path = RailsTemplate18f::Generators.const_source_location(name).first
+      end
+
+      class_methods do
+        attr_accessor :source_path
+
+        def source_root
+          @source_root ||= File.expand_path("templates", File.dirname(source_path))
+        end
+      end
+
+      private
+
+      def ruby_version
+        RUBY_VERSION
+      end
+
+      def skip_git?
+        !Dir.exist?(File.expand_path(".git", destination_root))
+      end
+    end
+  end
+end